// lib/api-zod/src/generated/types/paymentMode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalOrderHub - Multi-tenant local marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMode = typeof PaymentMode[keyof typeof PaymentMode];


export const PaymentMode = {
  ONLINE_ONLY: 'ONLINE_ONLY',
  PAY_AT_PICKUP_ONLY: 'PAY_AT_PICKUP_ONLY',
  BOTH: 'BOTH',
} as const;